tests/wpt/tests/websockets/interfaces/WebSocket/url/003.html

<!doctype html>
<title>WebSockets: deleting url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  delete ws.url;
  assert_equals(ws.url, SCHEME_DOMAIN_PORT+'/', 'delete ws.url');
  delete WebSocket.prototype.url;
  assert_equals(ws.url, undefined, 'delete WebSocket.prototype.url');
});
</script>
